requirements.txt
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.1.0
tqdm
colorama